# accounting/openapi/openapi.json
{
...
  "host": "127.0.0.1:45881",
  "info": {
    "title": "Komunitin accounting API",
    "version": "0.0.1",
    "license": {
      "name": "AGPL-3.0-only"
    },
    "description": "API for the Komunitin payments system. This documentation is automatically generated from the serivce test suite."
  },
  "paths": {
...
    "/{code}/cc/nodes": {
      "post": {
        "summary": "/{code}/cc/nodes",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/{code}/cc/": {
      "get": {
        "summary": "/{code}/cc/",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/{code}/cc/transaction/relay": {
      "post": {
        "summary": "/{code}/cc/transaction/relay",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/{code}/cc/transaction/{transId}/{newState}": {
      "patch": {
        "summary": "/{code}/cc/transaction/{transId}/{newState}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          },
          {
            "name": "transId",
            "in": "path",
            "required": true
          },
          {
            "name": "newState",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/{code}/cc/account": {
      "get": {
        "summary": "/{code}/cc/account",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/{code}/cc/account/history": {
      "get": {
        "summary": "/{code}/cc/account/history",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/{code}/cc/addresses/{id}": {
      "get": {
        "summary": "/{code}/cc/addresses/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
...
  },
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "authorization": {
      "name": "Authorization",
      "in": "header",
      "type": "apiKey",
      "description": "Bearer JWT from the configured Auth issuer for audience urn:komunitin:api. Use accounting:read for reads, accounting:write for writes, or superadmin on admin operations."
    }
  },
  "servers": [
    {
      "url": "https://accounting.komunitin.org",
      "description": "Production server"
    },
    {
      "url": "https://accounting.demo.komunitin.org",
      "description": "Demo server"
    },
    {
      "url": "http://localhost:2025",
      "description": "Local development server"
    }
  ],
  "swagger": "2.0",
...
}